refactor: MemberRoleMenu 리팩토링
