+++ components/members/main/MemberList/MemberRoleMenu/MemberRoleSelect.tsx
@@ -2,63 +2,20 @@ import styled from '@emotion/styled';
 import React, { FC } from 'react';
 
 import Select from '@/components/members/common/select/Select';
+import { MENUS } from '@/components/members/main/MemberList/MemberRoleMenu/constants';
 import { MOBILE_MEDIA_QUERY } from '@/styles/mediaQuery';
 import { textStyles } from '@/styles/typography';
 
-export const menuValue = {
-  ALL: '0',
-  PM: '1',
-  DESIGN: '2',
-  WEB: '3',
-  SERVER: '4',
-  ANDROID: '5',
-  iOS: '6',
-} as const;
-
-export const MENUS = [
-  {
-    icon: '/icons/icon-all.svg',
-    label: '전체',
-    value: menuValue.ALL,
-  },
-  {
-    icon: '/icons/icon-pm.svg',
-    label: 'PM',
-    value: menuValue.PM,
-  },
-  {
-    icon: '/icons/icon-design.svg',
-    label: '디자인',
-    value: menuValue.DESIGN,
-  },
-  {
-    icon: '/icons/icon-webpart.svg',
-    label: 'WEB',
-    value: menuValue.WEB,
-  },
-  {
-    icon: '/icons/icon-server.svg',
-    label: 'SERVER',
-    value: menuValue.SERVER,
-  },
-  {
-    icon: '/icons/icon-android.svg',
-    label: 'Android',
-    value: menuValue.ANDROID,
-  },
-  {
-    icon: '/icons/icon-iOS.svg',
-    label: 'iOS',
-    value: menuValue.iOS,
-  },
-];
 interface MemberRoleSelectProps {
   className?: string;
   value: string;
   onChange: (value: string) => void;
 }
-
-const MemberRoleSelect: FC<MemberRoleSelectProps> = ({ className, value, onChange }) => {
+/**
+ *
+ * @description 해당 컴포넌트는 /members 페이지의 모바일 반응형 화면에서만 사용됩니다.
+ */
+export const MemberRoleSelect: FC<MemberRoleSelectProps> = ({ className, value, onChange }) => {
   return (
     <StyledSelect className={className} value={value} onChange={onChange}>
       {MENUS.map((menu) => (
@@ -72,8 +29,6 @@ const MemberRoleSelect: FC<MemberRoleSelectProps> = ({ className, value, onChang
     </StyledSelect>
   );
 };
-
-export default MemberRoleSelect;
 
 const StyledSelect = styled(Select)`
   @media ${MOBILE_MEDIA_QUERY} {
